[r8379] Implemented attaching/detaching of external non-A2 threads in a DLL.
Every time a thread is created within the process which loaded the DLL, a DLL_THREAD_ATTACH event is handled in Objects.AttachThread called from Kernel32.EntryPoint (DllMain).
In Objects.AttachThread an attached external thread is put into the "ready" process queue and marked with a flag "Objects.External" to distinguish external threads from the ones created within A2.
Note, that the "stackBottom" field of all external processes is assigned to the HighLimit value obtained using Kernel32.GetCurrentThreadStackLimits. Note, that the thread which loaded the DLL
is also registered as an external one.

When an external thread is exiting a DLL_THREAD_DETACH event is handled in Objects.DetachThread. The thread to detach is removed from the process queue and all its allocated resources 
(e.g. event object used for synchronization) are cleaned up. When the DLL loading process is exiting a DLL_PROCESS_DETACH event is handled in Objects.DetachProcess, where Modules.Shutdown 
is executed and the processes Clock and FinalizerCaller are gracefully terminated (for that Exit method was implemented for both objects). Taking into account that termination of threads 
is a parallel process which could take some time, the code in Objects.DetachProcess waits until all threads terminate with subject to a timeout (currently 1 sec). All A2 threads which 
did not terminate in time are terminated forcedly using Kernel32.TerminateThread. After that the DLL TLS storage is freed.